Server of the dv.edish module

edish_server() contains the server of the dv.edish module.

Usage

edish_server(
  module_id,
  dataset_list,
  subjectid_var = "USUBJID",
  arm_var = "ACTARM",
  arm_default_vals = NULL,
  visit_var = "VISIT",
  baseline_visit_val = "VISIT 01",
  lb_test_var = "LBTEST",
  lb_test_choices = c("Alkaline Phosphatase", "Alanine Aminotransferase",
    "Aspartate Aminotransferase", "Bilirubin"),
  lb_test_default_x_val = "Aspartate Aminotransferase",
  lb_test_default_y_val = "Bilirubin",
  lb_result_var = "LBSTRESN",
  ref_range_upper_lim_var = "LBSTNRHI"
)

Arguments

module_id

[character(1)]

A unique ID string to create a namespace. Must match the ID of edish_UI().

dataset_list

[shiny::reactive(list(data.frame))]

A reactive list of named datasets.

subjectid_var

[character(1)]

Name of the variable containing the unique subject IDs.

arm_var

[character(1)]

Name of the variable containing the arm/treatment information.

arm_default_vals

[character(1+)]

Vector specifying the default value(s) for the arm selector.

visit_var

[character(1)]

Name of the variable containing the visit information.

baseline_visit_val

[character(1)]

Character indicating which visit should be used as baseline visit.

lb_test_var

[character(1)]

Name of the variable containing the laboratory test information.

lb_test_choices

[character(1+)]

Character vector specifying the possible choices of the laboratory test.

lb_test_default_x_val

[character(1)]

Character specifying the default laboratory test choice for the plot's x-axis.

lb_test_default_y_val

[character(1)]

Character specifying the default laboratory test choice for the plot's y-axis.

lb_result_var

[character(1)]

Name of the variable containing results of the laboratory test.

ref_range_upper_lim_var

[character(1)]

Name of the variable containing the reference range upper limits.

See also