Generate plot

generate_plot() generates an eDISH plot by means of the plotly package.

Usage

generate_plot(
  dataset,
  subjectid_var,
  arm_var,
  visit_var,
  sel_x,
  sel_y,
  x_plot_type,
  y_plot_type,
  x_ref_line_num,
  y_ref_line_num,
  x_rng_lower,
  x_rng_upper,
  y_rng_lower,
  y_rng_upper,
  source = NULL
)

Arguments

dataset

[data.frame]

A data frame containing the variables listed below as columns.

subjectid_var

[character(1)]

Name of the variable containing the unique subject IDs.

arm_var

[character(1)]

Name of the variable containing the arm/treatment information.

visit_var

[character(1)]

Name of the variable containing the visit information.

sel_x

[character(1)]

Character specifying the laboratory test to be displayed on the x-axis.

sel_y

[character(1)]

Character specifying the laboratory test to be displayed on the y-axis.

x_plot_type

[character(1)]

Character specifying the plot type for the x-axis. This leads to using the dataset's column "r_<x_plot_type>_<x_sel>" for the x-values.

y_plot_type

[character(1)]

Character specifying the plot type for the y-axis. This leads to using the dataset's column "r_<y_plot_type>_<y_sel>" for the y-values.

x_ref_line_num

[numeric(1)]

Numeric specifying the reference line for the x-axis.

y_ref_line_num

[numeric(1)]

Numeric specifying the reference line for the y-axis.

x_rng_lower

[numeric(1)]

Numeric specifying the lower limit in the x-axis range.

x_rng_upper

[numeric(1)]

Numeric specifying the upper limit in the x-axis range.

y_rng_lower

[numeric(1)]

Numeric specifying the lower limit in the y-axis range.

y_rng_upper

[numeric(1)]

Numeric specifying the upper limit in the y-axis range.

Value

A plotly object specifying the generated eDISH plot.